# src/swagger.yaml
openapi: 3.0.0
info:
  title: AI Model API
  description: An API for generating AI model responses.
  version: 1.0.0
servers:
- url: /
paths:
  /stream:
    get:
      summary: Generates AI model response
      description: This route generates an AI model response using the given prompt, model choice, and maximum tokens.
      parameters:
      - name: prompt
        in: query
        description: The prompt to generate the AI model response.
        required: true
        schema:
          type: string
      - name: modelChoice
        in: query
        description: The AI model choice (default is 'gpt-3.5-turbo').
        required: false
        schema:
          type: string
      - name: maxTokens
        in: query
        description: The maximum number of tokens for the generated response (default is 1000, cannot exceed 4096).
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Successful generation of AI model response.
          content:
            text/stream:
              schema:
                type: string
        '400':
          description: Bad request, invalid input provided or exceeded the allowed token limit.
          content:
            text/plain:
              schema:
                type: string
  /stream/chat:
    get:
      summary: Generates chat model response
      description: This route generates a chat model response using the given prompt, model choice, and maximum tokens.
      parameters:
      - name: prompt
        in: query
        description: The prompt to generate the chat model response.
        required: true
        schema:
          type: string
      - name: modelChoice
        in: query
        description: The chat model choice (default is 'gpt-3.5-turbo').
        required: false
        schema:
          type: string
      - name: maxTokens
        in: query
        description: The maximum number of tokens for the generated response (default is 1000, cannot exceed 4096).
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Successful generation of chat model response.
          content:
            text/stream:
              schema:
                type: string
        '400':
          description: Bad request, invalid input provided or exceeded the allowed token limit.
          content:
            text/plain:
              schema:
                type: string
...